Refactor Udacity_OpenSBTWrapper class and import statements
