+++ examples/sdsandbox_perturbations/main.py
@@ -2,8 +2,6 @@
 from problem.adas_problem import ADASProblem
 from evaluation.fitness import *
 from evaluation.critical import *
-from simulation.simulator import Simulator, SimulationOutput
-from model_ga.individual import Individual
 from algorithm.nsga2_optimizer import NsgaIIOptimizer
 from experiment.search_configuration import DefaultSearchConfiguration
 
@@ -14,175 +12,14 @@ import traceback
 
 from sdsandbox_simulator import SDSandboxSimulator
 from examples.models.example_agent import ExampleAgent
-from examples.open_sbt.Criticality import FitnessFunction, Criticality
+from examples.open_sbt.criticality import FitnessFunction, Criticality
+from examples.open_sbt.sdsandbox_open_sbt import SDSandBox_OpenSBTWrapper
 
 # related to perturbation drive
 from perturbationdrive import (
     PerturbationDrive,
     RandomRoadGenerator,
-    ScenarioOutcome,
-    Scenario,
-    CustomRoadGenerator,
 )
-
-
-class SDSandBox_OpenSBTWrapper(Simulator):
-    @staticmethod
-    def simulate(
-        list_individuals: List[Individual],
-        variable_names: List[str],
-        scenario_path: str,
-        sim_time: float,
-        time_step: float,
-        do_visualize: bool = False,
-    ) -> List[SimulationOutput]:
-        """
-        Runs all indicidual simulations and returns simulation outputs for each individual
-        """
-        # set up all perturbation drive objects
-        simulator = SDSandboxSimulator(
-            simulator_exe_path="./examples/sdsandbox_perturbations/sim/donkey-sim.app",
-            host="127.0.0.1",
-            port=9091,
-        )
-        ads = ExampleAgent()
-        benchmarking_obj = PerturbationDrive(simulator, ads)
-        road_generator = CustomRoadGenerator(250)
-
-        # we need to set the sim here up to get the starting position
-        benchmarking_obj.simulator.connect()
-        starting_pos = benchmarking_obj.simulator.initial_pos
-
-        # create all scenarios
-        scenarios: List[Scenario] = [
-            SDSandBox_OpenSBTWrapper.individualToScenario(
-                individual=ind,
-                variable_names=variable_names,
-                road_generator=road_generator,
-                starting_pos=starting_pos,
-            )
-            for ind in list_individuals
-        ]
-
-        # run the individualts
-        outcomes: List[ScenarioOutcome] = benchmarking_obj.simulate_scenarios(
-            scenarios=scenarios,
-            attention_map={},
-            log_dir=None,
-            overwrite_logs=False,
-            image_size=(240, 320),
-        )
-
-        # convert the outcomes to sbt format
-        return [
-            SimulationOutput(
-                simTime=float(len(outcome.frames)),
-                times=outcome.frames,
-                location={"ego": [(x[0], x[1]) for x in outcome.pos]},
-                velocity={
-                    "ego": SDSandBox_OpenSBTWrapper._calculate_velocities(
-                        outcome.pos, outcome.speeds
-                    )
-                },
-                speed={"ego": outcome.speeds},
-                acceleration={"ego": []},
-                yaw={
-                    "ego": [],
-                },
-                collisions=[],
-                actors={
-                    1: "ego",
-                },
-                otherParams={"xte": outcome.xte},
-            )
-            for outcome in outcomes
-        ]
-
-    @staticmethod
-    def individualToScenario(
-        individual: Individual,
-        variable_names: List[str],
-        road_generator: CustomRoadGenerator,
-        starting_pos: Tuple[float, float, float],
-    ) -> Scenario:
-        instance_values = [v for v in zip(variable_names, individual)]
-        angles: List[str] = []
-        perturbation_scale: int = 0
-        perturbation_function_int: int = 1
-        perturbation_function: str = ""
-        seg_lengths: List[str] = []
-        for i in range(0, len(instance_values) + 1):
-            # Check if the current item is the perturbation scale
-            if instance_values[i][0].startswith("perturbation_scale"):
-                perturbation_scale = int(instance_values[i][1])
-                break
-            elif instance_values[i][0].startswith("perturbation_function"):
-                perturbation_function_int = int(instance_values[i][1])
-                break
-            elif instance_values[i][0].startswith("angle"):
-                new_angle = int(instance_values[i][1])
-                angles.append(new_angle)
-            elif instance_values[i][0].startswith("seg_length"):
-                seg_length = int(instance_values[i][1])
-                seg_lengths.append(seg_length)
-
-        # generate the road string from the configuration
-        seg_lengths: Union[List[str], None] = (
-            seg_lengths if len(seg_lengths) > 0 else None
-        )
-        road_str: str = road_generator.generate(
-            starting_pos=starting_pos, angles=angles, seg_lengths=seg_lengths
-        )
-        # map the function
-        amount_keys = len(list(FUNCTION_MAPPING.keys()))
-        if perturbation_function_int > 0 and perturbation_function_int <= amount_keys:
-            perturbation_function = FUNCTION_MAPPING[perturbation_function_int]
-        else:
-            perturbation_function = FUNCTION_MAPPING[1]
-            print(
-                f"Perturbation function not found for values {perturbation_function_int}, using default"
-            )
-
-        # return the scenario
-        return Scenario(
-            waypoints=road_str,
-            perturbation_function=perturbation_function,
-            perturbation_scale=perturbation_scale,
-        )
-
-    @staticmethod
-    def _calculate_velocities(
-        positions: List[Tuple[float, float, float]], speeds: List[float]
-    ) -> Tuple[float, float, float]:
-        """
-        Calculate velocities given a list of positions and corresponding speeds.
-        """
-        if len(positions) != len(speeds) or len(speeds) <= 1:
-            return []
-        velocities = []
-        for i in range(len(positions) - 1):
-            displacement = np.array(positions[i + 1]) - np.array(positions[i])
-            displacement_norm = np.linalg.norm(displacement)
-            # avoid division by zero
-            if displacement_norm > 0:
-                displacement_norm += 0.001
-            else:
-                displacement_norm -= 0.001
-            direction = displacement / displacement_norm
-            velocity = direction * speeds[i]
-            velocities.append(velocity)
-
-        return velocities
-
-
-FUNCTION_MAPPING = {
-    1: "gaussian_noise",
-    2: "poisson_noise",
-    3: "impulse_noise",
-    4: "defocus_blur",
-    5: "glass_blur",
-    6: "increase_brightness",
-}
 
 
 def open_sbt():
